Added new Class: Collectable_list and levitate method onto Collectables class
Instead of having a list variable, I've made it into an class object.
After a lot of testing, I found out that the server kept overwriting the collectable item data, so any change from the client side could not be processed. Next step is to move most of the processing to the client side and leave only the instantiation on the server side.